requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
customtkinter>=5.2.0
plotly>=5.14.0
pandas>=2.0.0